#9 Refactor auth to authentication and create authorization for separation of concerns.
