+++ backend/routes/carts-routes.js
@@ -1,17 +1,17 @@
 const express = require("express");
 const router = express.Router();
 const cartController = require("../controllers/cart-controller");
-const authMiddleware = require("../middlewares/auth");
+const authentication = require("../middlewares/authentication");
 
-router.get("/", authMiddleware.authenticate, cartController.getCart);
+router.get("/", authentication.authenticateUser, cartController.getCart);
 
-router.put("/:id", authMiddleware.authenticate, cartController.updateCart);
+router.put("/:id", authentication.authenticateUser, cartController.updateCart);
 
-router.post("/", authMiddleware.authenticate, cartController.createCart);
+router.post("/", authentication.authenticateUser, cartController.createCart);
 
 router.delete(
   "/:id",
-  authMiddleware.authenticate,
+  authentication.authenticateUser,
   cartController.deleteCart
 );
