+++ backend/routes/discounts-routes.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const discountController = require("../controllers/discount-controller");
-const authMiddleware = require("../middlewares/auth");
+const authentication = require("../middlewares/authentication");
 
-router.get("/", authMiddleware.authenticate, cartController.getCategories);
+router.get("/", authentication.authenticateUser, cartController.getCategories);
 
 module.exports = router;
